requirements.txt
appdirs==1.4.4
backcall==0.2.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
decorator==4.4.2
Flask==1.1.2
gunicorn==20.1.0
idna==2.10
ipykernel==5.5.3
ipython==7.22.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
jupyter-client==6.1.12
jupyter-core==4.7.1
MarkupSafe==1.1.1
numpy==1.20.2
opencv-contrib-python==4.5.1.48
parso==0.8.1
pickleshare==0.7.5
Pillow==8.1.2
prompt-toolkit==3.0.18
Pygments==2.8.1
python-dateutil==2.8.1
pyzmq==22.0.3
requests==2.25.1
six==1.15.0
torch==1.8.1
torchaudio==0.8.1
torchvision==0.9.1
tornado==6.1
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.4
wcwidth==0.2.5
Werkzeug==1.0.1
